+++ packages/react-relay/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.8...@wundergraph/react-relay@0.4.9) (2023-08-17)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.4.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.7...@wundergraph/react-relay@0.4.8) (2023-08-14)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.4.7](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.6...@wundergraph/react-relay@0.4.7) (2023-08-13)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.4.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.5...@wundergraph/react-relay@0.4.6) (2023-08-03)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.4.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.4...@wundergraph/react-relay@0.4.5) (2023-08-03)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.4.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.3...@wundergraph/react-relay@0.4.4) (2023-08-02)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.4.3](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.2...@wundergraph/react-relay@0.4.3) (2023-08-01)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.4.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.1...@wundergraph/react-relay@0.4.2) (2023-07-28)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
 ## [0.4.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.4.0...@wundergraph/react-relay@0.4.1) (2023-07-24)
 
 ### Bug Fixes